Cretaes the move robot class
